Fix bug in ChunkType and add try_from
Fix the bug in ChunkType's try_from and from_str by changing the
lower_range from 67 to 65 which was a mistake. Add try_from method full
implementation to store the chunks and a header in the Png struct and
pass necessary tests. Add from_chunks method to represent a PNG file
from given chunks.